Make only page, external and relative links available on base link block
