+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/common/base_link_block.py
@@ -4,7 +4,7 @@ from django.utils.functional import cached_property
 from wagtail import blocks
 from wagtail.blocks.struct_block import StructBlockAdapter
 
-from networkapi.wagtailpages.validators import AnchorLinkValidator, RelativeURLValidator
+from networkapi.wagtailpages.validators import RelativeURLValidator
 
 
 class BaseLinkValue(blocks.StructValue):
@@ -21,12 +21,6 @@ class BaseLinkValue(blocks.StructValue):
 
     def get_relative_url_link(self):
         return self.get("relative_url")
-
-    def get_email_link(self):
-        return f"mailto:{self.get('email')}"
-
-    def get_anchor_link(self):
-        return f"#{self.get('anchor')}"
 
     @property
     def url(self):
@@ -48,8 +42,6 @@ class BaseLinkBlock(blocks.StructBlock):
             ("page", "Page"),
             ("external_url", "External URL"),
             ("relative_url", "Relative URL"),
-            ("email", "Email"),
-            ("anchor", "Anchor"),
         ],
         required=False,
         label="Link to",
@@ -68,14 +60,6 @@ class BaseLinkBlock(blocks.StructBlock):
         label="Relative URL",
         help_text='A path relative to this domain. For example, "/foo/bar"',
     )
-    anchor = blocks.CharBlock(
-        max_length=300,
-        required=False,
-        validators=[AnchorLinkValidator()],
-        label="#",
-        help_text='An id attribute of an element on the current page. For example, "#section-1"',
-    )
-    email = blocks.EmailBlock(required=False)
 
     class Meta:
         abstract = True
@@ -86,8 +70,6 @@ class BaseLinkBlock(blocks.StructBlock):
             "page": None,
             "external_url": "",
             "relative_url": "",
-            "anchor": "",
-            "email": "",
         }
 
     def clean(self, value):
@@ -96,6 +78,9 @@ class BaseLinkBlock(blocks.StructBlock):
 
         url_default_values = self.get_default_values()
         url_type = clean_values.get("link_to")
+
+        if not url_type:
+            errors["link_to"] = ErrorList(["Please select a link type"])
 
         # Check that a value has been uploaded for the chosen link type
         if url_type != "" and clean_values.get(url_type) in [None, ""]:
